Fix dashboard overlay eagerly requesting global user presence
