+++ osu.Game/Overlays/Dashboard/CurrentlyOnline/CurrentlyOnlineDisplay.cs
@@ -1,13 +1,16 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System;
 using System.Linq;
 using System.Threading;
 using osu.Framework.Allocation;
+using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics.UserInterface;
+using osu.Game.Online.Metadata;
 using osu.Game.Overlays.Dashboard.Friends;
 using osu.Game.Resources.Localisation.Web;
 
@@ -15,6 +18,17 @@ namespace osu.Game.Overlays.Dashboard.CurrentlyOnline
 {
     public partial class CurrentlyOnlineDisplay : CompositeDrawable
     {
+        /// <summary>
+        /// The current state of the <see cref="DashboardOverlay"/>.
+        /// Presence is only updated when this value is <see cref="Visibility.Visible"/>.
+        /// </summary>
+        public readonly Bindable<Visibility> OverlayState = new Bindable<Visibility>(Visibility.Visible);
+
+        [Resolved]
+        private MetadataClient metadataClient { get; set; } = null!;
+
+        private readonly IBindable<bool> isConnected = new Bindable<bool>();
+
         private Box background = null!;
         private UserListToolbar userListToolbar = null!;
         private Container<RealtimeUserList> listContainer = null!;
@@ -22,6 +36,7 @@ namespace osu.Game.Overlays.Dashboard.CurrentlyOnline
         private BasicSearchTextBox searchTextBox = null!;
 
         private CancellationTokenSource? listLoadCancellation;
+        private IDisposable? userPresenceWatchToken;
 
         public CurrentlyOnlineDisplay()
         {
@@ -114,6 +129,11 @@ namespace osu.Game.Overlays.Dashboard.CurrentlyOnline
         {
             base.LoadComplete();
 
+            isConnected.BindTo(metadataClient.IsConnected);
+            isConnected.BindValueChanged(_ => updateUserPresenceState());
+
+            OverlayState.BindValueChanged(_ => updateUserPresenceState(), true);
+
             userListToolbar.DisplayStyle.BindValueChanged(_ => reloadList(), true);
         }
 
@@ -147,12 +167,28 @@ namespace osu.Game.Overlays.Dashboard.CurrentlyOnline
             }
         }
 
+        private void updateUserPresenceState()
+        {
+            if (!isConnected.Value)
+                return;
+
+            if (OverlayState.Value == Visibility.Visible)
+                userPresenceWatchToken ??= metadataClient.BeginWatchingUserPresence();
+            else
+            {
+                userPresenceWatchToken?.Dispose();
+                userPresenceWatchToken = null;
+            }
+        }
+
         protected override void Dispose(bool isDisposing)
         {
             base.Dispose(isDisposing);
 
             listLoadCancellation?.Cancel();
             listLoadCancellation?.Dispose();
+
+            userPresenceWatchToken?.Dispose();
         }
     }
 }
