+++ osu.Game/Overlays/DashboardOverlay.cs
@@ -2,10 +2,6 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using osu.Framework.Allocation;
-using osu.Framework.Bindables;
-using osu.Framework.Graphics.Containers;
-using osu.Game.Online.Metadata;
 using osu.Game.Overlays.Dashboard;
 using osu.Game.Overlays.Dashboard.CurrentlyOnline;
 using osu.Game.Overlays.Dashboard.Friends;
@@ -14,12 +10,6 @@ namespace osu.Game.Overlays
 {
     public partial class DashboardOverlay : TabbableOnlineOverlay<DashboardOverlayHeader, DashboardOverlayTabs>
     {
-        [Resolved]
-        private MetadataClient metadataClient { get; set; } = null!;
-
-        private IBindable<bool> metadataConnected = null!;
-        private IDisposable? userPresenceWatchToken;
-
         public DashboardOverlay()
             : base(OverlayColourScheme.Purple)
         {
@@ -38,35 +28,14 @@ namespace osu.Game.Overlays
                     break;
 
                 case DashboardOverlayTabs.CurrentlyPlaying:
-                    LoadDisplay(new CurrentlyOnlineDisplay());
+                    LoadDisplay(new CurrentlyOnlineDisplay
+                    {
+                        OverlayState = { BindTarget = State }
+                    });
                     break;
 
                 default:
                     throw new NotImplementedException($"Display for {tab} tab is not implemented");
-            }
-        }
-
-        protected override void LoadComplete()
-        {
-            base.LoadComplete();
-
-            metadataConnected = metadataClient.IsConnected.GetBoundCopy();
-            metadataConnected.BindValueChanged(_ => updateUserPresenceState());
-            State.BindValueChanged(_ => updateUserPresenceState());
-            updateUserPresenceState();
-        }
-
-        private void updateUserPresenceState()
-        {
-            if (!metadataConnected.Value)
-                return;
-
-            if (State.Value == Visibility.Visible)
-                userPresenceWatchToken ??= metadataClient.BeginWatchingUserPresence();
-            else
-            {
-                userPresenceWatchToken?.Dispose();
-                userPresenceWatchToken = null;
             }
         }
     }
